--- a/lib/concurrent.js
+++ b/lib/concurrent.js
@@ -7,7 +7,7 @@ module.exports = Concurrent;
 
 function Concurrent(tests, bail) {
 	if (!this instanceof Concurrent) {
-		throw new Error('must use `new Concurrent`');
+		throw new TypeError('Class constructor Concurrent cannot be invoked without \'new\'');
 	}
 	this.tests = tests;
 	this.bail = bail;
@@ -74,9 +74,9 @@ Concurrent.prototype.run = function () {
 	});
 
 	if (bail) {
-		ret = ret.catch(function (e) {
-			if (e !== BAIL_ERROR) {
-				throw e;
+		ret = ret.catch(function (err) {
+			if (err !== BAIL_ERROR) {
+				throw err;
 			}
 			return {
 				passed: false,