--- a/lib/sequence.js
+++ b/lib/sequence.js
@@ -6,7 +6,7 @@ module.exports = Sequence;
 
 function Sequence(tests, bail) {
 	if (!this instanceof Sequence) {
-		throw new Error('Sequence must be called with new');
+		throw new TypeError('Class constructor Sequence cannot be invoked without \'new\'');
 	}
 	this.tests = tests;
 	this.bail = bail;