--- a/lib/test-collection.js
+++ b/lib/test-collection.js
@@ -34,3 +34,79 @@ TestCollection.prototype.add = function (test) {
 	}
 	this.tests[type].push(test);
 };
+
+var fnName = require('fn-name');
+var Sequence = require('./sequence');
+var Concurrent = require('./concurrent');
+var Test = require('./test');
+
+function computeTitle(entry) {
+	entry.title = entry.title || fnName(entry.fn);
+
+	// workaround for Babel giving anonymous functions a name
+	if (entry.title === 'callee$0$0') {
+		entry.title = null;
+	}
+
+	if (!entry.title) {
+		if (entry.metadata.type === 'test') {
+			entry.title = '[anonymous]';
+		} else {
+			entry.title = entry.metadata.type;
+		}
+	}
+}
+
+function buildHooks(hookArray, title, contextRef) {
+	return hookArray.map(function (hook) {
+		var test = new Test(hook.title + ' for ' + title, hook.fn, contextRef);
+		test.metadata = hook.metadata;
+		return test;
+	});
+}
+
+function buildTest(entry, ctxRef) {
+	var test = new Test(entry.title, entry.fn, ctxRef);
+	test.metadata = entry.metadata;
+	return test;
+}
+
+function buildTestWithHooks(entry, beforeEach, afterEach) {
+	var contextRef = {context: {}};
+	var arr = buildHooks(beforeEach, entry.title, contextRef);
+	arr.push(buildTest(entry, contextRef));
+	return arr.concat(buildHooks(afterEach, entry.title, contextRef));
+}
+
+TestCollection.prototype.buildPhases = function () {
+	[
+		this.serial,
+		this.concurrent,
+		this.tests.before,
+		this.tests.beforeEach,
+		this.tests.after,
+		this.tests.afterEach
+	].forEach(function (arr) {
+		arr.forEach(computeTitle);
+	});
+
+	return new Sequence(
+		[
+			new Sequence(this.tests.before.map(function (entry) {
+				return buildTest(entry);
+			})),
+			new Sequence([
+				new Sequence(this.serial.map(function (entry) {
+					return new Sequence(buildTestWithHooks(entry, this.tests.beforeEach, this.tests.afterEach), true);
+				}, this), false),
+				new Concurrent(this.concurrent.map(function (entry) {
+					return new Sequence(buildTestWithHooks(entry, this.tests.beforeEach, this.tests.afterEach), true);
+				}, this), false)
+			], false),
+			new Sequence(this.tests.after.map(function (entry) {
+				return buildTest(entry);
+			}))
+		],
+		true
+	);
+};