--- a/lib/test-collection.js
+++ b/lib/test-collection.js
@@ -57,28 +57,28 @@ function computeTitle(entry) {
 	}
 }
 
-function buildHooks(hookArray, title, contextRef) {
+function buildHooks(hookArray, title, contextRef, report) {
 	return hookArray.map(function (hook) {
-		var test = new Test(hook.title + ' for ' + title, hook.fn, contextRef);
+		var test = new Test(hook.title + ' for ' + title, hook.fn, contextRef, report);
 		test.metadata = hook.metadata;
 		return test;
 	});
 }
 
-function buildTest(entry, ctxRef) {
-	var test = new Test(entry.title, entry.fn, ctxRef);
+function buildTest(entry, ctxRef, report) {
+	var test = new Test(entry.title, entry.fn, ctxRef, report);
 	test.metadata = entry.metadata;
 	return test;
 }
 
-function buildTestWithHooks(entry, beforeEach, afterEach) {
+function buildTestWithHooks(entry, beforeEach, afterEach, report) {
 	var contextRef = {context: {}};
-	var arr = buildHooks(beforeEach, entry.title, contextRef);
-	arr.push(buildTest(entry, contextRef));
-	return arr.concat(buildHooks(afterEach, entry.title, contextRef));
+	var arr = buildHooks(beforeEach, entry.title, contextRef, report);
+	arr.push(buildTest(entry, contextRef, report));
+	return arr.concat(buildHooks(afterEach, entry.title, contextRef, report));
 }
 
-TestCollection.prototype.buildPhases = function () {
+TestCollection.prototype.buildPhases = function (report) {
 	[
 		this.serial,
 		this.concurrent,
@@ -93,18 +93,18 @@ TestCollection.prototype.buildPhases = function () {
 	return new Sequence(
 		[
 			new Sequence(this.tests.before.map(function (entry) {
-				return buildTest(entry);
+				return buildTest(entry, null, report);
 			})),
 			new Sequence([
 				new Sequence(this.serial.map(function (entry) {
-					return new Sequence(buildTestWithHooks(entry, this.tests.beforeEach, this.tests.afterEach), true);
+					return new Sequence(buildTestWithHooks(entry, this.tests.beforeEach, this.tests.afterEach, report), true);
 				}, this), false),
 				new Concurrent(this.concurrent.map(function (entry) {
-					return new Sequence(buildTestWithHooks(entry, this.tests.beforeEach, this.tests.afterEach), true);
+					return new Sequence(buildTestWithHooks(entry, this.tests.beforeEach, this.tests.afterEach, report), true);
 				}, this), false)
 			], false),
 			new Sequence(this.tests.after.map(function (entry) {
-				return buildTest(entry);
+				return buildTest(entry, null, report);
 			}))
 		],
 		true