--- a/lib/test-collection.js
+++ b/lib/test-collection.js
@@ -23,8 +23,13 @@ TestCollection.prototype.add = function (test) {
 		throw new Error('test type must be specified');
 	}
 	if (type === 'test') {
-		if (metadata.exclusive) {
+		if (this.hasExclusive && !metadata.exclusive) {
+			return;
+		}
+		if (metadata.exclusive && !this.hasExclusive) {
 			this.hasExclusive = true;
+			this.serial = [];
+			this.concurrent = [];
 		}
 		(metadata.serial ? this.serial : this.concurrent).push(test);
 		return;
@@ -59,10 +64,18 @@ function computeTitle(entry) {
 
 function buildHooks(hookArray, title, contextRef, report) {
 	return hookArray.map(function (hook) {
-		var test = new Test(hook.title + ' for ' + title, hook.fn, contextRef, report);
-		test.metadata = hook.metadata;
+		var test = buildHook(hook, title, contextRef, report);
+		if (hook.metadata.skipped) {
+			return skipRunner(test, report);
+		}
 		return test;
 	});
+}
+
+function buildHook(hook, title, contextRef, report) {
+	var test = new Test(hook.title + ' for ' + title, hook.fn, contextRef, report);
+	test.metadata = hook.metadata;
+	return test;
 }
 
 function buildTest(entry, ctxRef, report) {
@@ -71,7 +84,25 @@ function buildTest(entry, ctxRef, report) {
 	return test;
 }
 
+function skipRunner(test, report) {
+	return {
+		run: function () {
+			var result = {
+				passed: true,
+				result: test
+			};
+			if (report) {
+				report(result);
+			}
+			return result;
+		}
+	};
+}
+
 function buildTestWithHooks(entry, beforeEach, afterEach, report) {
+	if (entry.metadata.skipped) {
+		return [skipRunner(buildTest(entry), report)];
+	}
 	var contextRef = {context: {}};
 	var arr = buildHooks(beforeEach, entry.title, contextRef, report);
 	arr.push(buildTest(entry, contextRef, report));
@@ -93,6 +124,9 @@ TestCollection.prototype.buildPhases = function (report) {
 	return new Sequence(
 		[
 			new Sequence(this.tests.before.map(function (entry) {
+				if (entry.metadata.skipped) {
+					return skipRunner(entry, report);
+				}
 				return buildTest(entry, null, report);
 			})),
 			new Sequence([
@@ -104,6 +138,9 @@ TestCollection.prototype.buildPhases = function (report) {
 				}, this), false)
 			], false),
 			new Sequence(this.tests.after.map(function (entry) {
+				if (entry.metadata.skipped) {
+					return skipRunner(entry, report);
+				}
 				return buildTest(entry, null, report);
 			}))
 		],