--- a/lib/test-collection.js
+++ b/lib/test-collection.js
@@ -35,7 +35,7 @@ TestCollection.prototype.add = function (test) {
 		return;
 	}
 	if (metadata.exclusive) {
-		throw new Error('you can\'t use only with a ' + type + ' test');
+		throw new Error('you can\'t use "only" with a ' + type + ' test');
 	}
 	this.tests[type].push(test);
 };