--- a/lib/test.js
+++ b/lib/test.js
@@ -29,6 +29,7 @@ function Test(title, fn) {
 	this.planCount = null;
 	this.duration = null;
 	this.assertError = undefined;
+	this.sync = true;
 
 	// TODO(jamestalmage): make this an optional constructor arg instead of having Runner set it after the fact.
 	// metadata should just always exist, otherwise it requires a bunch of ugly checks all over the place.
@@ -54,6 +55,9 @@ Object.defineProperty(Test.prototype, 'assertCount', {
 });
 
 Test.prototype._assert = function (promise) {
+	if (isPromise(promise)) {
+		this.sync = false;
+	}
 	this.assertions.push(promise);
 };
 
@@ -87,19 +91,28 @@ Test.prototype._run = function () {
 		ret = this.fn(this._publicApi());
 	} catch (err) {
 		this._setAssertError(err);
-		this.exit();
 	}
 	return ret;
 };
 
-Test.prototype.run = function () {
+Test.prototype.promise = function () {
 	var self = this;
+	if (!this._promise) {
+		this._promise = {};
+		this._promise.promise = new Promise(function (resolve, reject) {
+			self._promise.resolve = resolve;
+			self._promise.reject = reject;
+		});
+	}
+	return this._promise;
+};
 
-	this.promise = {};
-	this.promise.promise = new Promise(function (resolve, reject) {
-		self.promise.resolve = resolve;
-		self.promise.reject = reject;
-	});
+Test.prototype.run = function () {
+	if (this.metadata.callback) {
+		this.sync = false;
+	}
+
+	var self = this;
 
 	this._timeStart = globals.now();
 
@@ -116,10 +129,14 @@ Test.prototype.run = function () {
 	}
 
 	if (isPromise(ret)) {
+		this.sync = false;
+
 		if (this.metadata.callback) {
 			self._setAssertError(new Error('Do not return ' + asyncType + ' from tests declared via `test.cb(...)`, if you want to return a promise simply declare the test via `test(...)`'));
-			this.exit();
-			return this.promise.promise;
+		}
+
+		if (this.assertError !== undefined) {
+			return this.exit();
 		}
 
 		ret
@@ -138,11 +155,18 @@ Test.prototype.run = function () {
 				self._setAssertError(err);
 				self.exit();
 			});
-	} else if (!this.metadata.callback) {
-		this.exit();
 	}
 
-	return this.promise.promise;
+	if (!this.metadata.callback) {
+		return this.exit();
+	}
+
+	return this.promise().promise;
+};
+
+Test.prototype._result = function () {
+	var passed = this.assertError === undefined;
+	return {passed: passed, result: this, reason: this.assertError};
 };
 
 Object.defineProperty(Test.prototype, 'end', {
@@ -192,6 +216,12 @@ Test.prototype.exit = function () {
 
 	this._checkPlanCount();
 
+	if (this.sync) {
+		self.duration = globals.now() - self._timeStart;
+		globals.clearTimeout(self._timeout);
+		return this._result();
+	}
+
 	Promise.all(this.assertions)
 		.catch(function (err) {
 			self._setAssertError(err);
@@ -205,13 +235,10 @@ Test.prototype.exit = function () {
 
 			self._checkPlanCount();
 
-			if (self.assertError !== undefined) {
-				self.promise.reject(self.assertError);
-				return;
-			}
-
-			self.promise.resolve(self);
+			self.promise().resolve(self._result());
 		});
+
+	return self.promise().promise;
 };
 
 Test.prototype._publicApi = function () {
@@ -225,24 +252,23 @@ function PublicApi(test) {
 }
 
 function onAssertionEvent(event) {
-	var promise;
+	var promise = null;
 	if (event.assertionThrew) {
 		event.error.powerAssertContext = event.powerAssertContext;
 		event.error.originalMessage = event.originalMessage;
-		promise = Promise.reject(event.error);
+		this._test._setAssertError(event.error);
 	} else {
 		var ret = event.returnValue;
 		if (isObservable(ret)) {
 			ret = observableToPromise(ret);
 		}
 		if (isPromise(ret)) {
-			ret = ret
+			promise = ret
 				.then(null, function (err) {
 					err.originalMessage = event.originalMessage;
 					throw err;
 				});
 		}
-		promise = ret;
 	}
 	this._test._assert(promise);
 	return promise;
@@ -281,7 +307,7 @@ Object.defineProperty(PublicApi.prototype, 'context', {
 });
 
 function skipFn() {
-	return this._test._assert(Promise.resolve());
+	return this._test._assert(null);
 }
 
 function SkipApi(test) {