--- a/lib/test.js
+++ b/lib/test.js
@@ -11,12 +11,13 @@ var assert = require('./assert');
 var enhanceAssert = require('./enhance-assert');
 var globals = require('./globals');
 
-function Test(title, fn) {
+function Test(title, fn, contextRef) {
 	if (!(this instanceof Test)) {
-		return new Test(title, fn);
+		return new Test(title, fn, contextRef);
 	}
 
 	if (typeof title === 'function') {
+		contextRef = fn;
 		fn = title;
 		title = null;
 	}
@@ -30,6 +31,7 @@ function Test(title, fn) {
 	this.duration = null;
 	this.assertError = undefined;
 	this.sync = true;
+	this.contextRef = contextRef;
 
 	// TODO(jamestalmage): make this an optional constructor arg instead of having Runner set it after the fact.
 	// metadata should just always exist, otherwise it requires a bunch of ugly checks all over the place.
@@ -299,10 +301,10 @@ PublicApi.prototype = enhanceAssert({
 Object.defineProperty(PublicApi.prototype, 'context', {
 	enumerable: true,
 	get: function () {
-		return this._test.context;
+		return this._test.contextRef.context;
 	},
 	set: function (context) {
-		this._test.context = context;
+		this._test.contextRef.context = context;
 	}
 });