--- a/lib/test.js
+++ b/lib/test.js
@@ -11,7 +11,7 @@ var assert = require('./assert');
 var enhanceAssert = require('./enhance-assert');
 var globals = require('./globals');
 
-function Test(title, fn, contextRef) {
+function Test(title, fn, contextRef, report) {
 	if (!(this instanceof Test)) {
 		return new Test(title, fn, contextRef);
 	}
@@ -32,6 +32,7 @@ function Test(title, fn, contextRef) {
 	this.assertError = undefined;
 	this.sync = true;
 	this.contextRef = contextRef;
+	this.report = report;
 
 	// TODO(jamestalmage): make this an optional constructor arg instead of having Runner set it after the fact.
 	// metadata should just always exist, otherwise it requires a bunch of ugly checks all over the place.
@@ -104,6 +105,10 @@ Test.prototype.promise = function () {
 		this._promise.promise = new Promise(function (resolve, reject) {
 			self._promise.resolve = resolve;
 			self._promise.reject = reject;
+		}).tap(function (result) {
+			if (self.report) {
+				self.report(result);
+			}
 		});
 	}
 	return this._promise;
@@ -137,15 +142,11 @@ Test.prototype.run = function () {
 			self._setAssertError(new Error('Do not return ' + asyncType + ' from tests declared via `test.cb(...)`, if you want to return a promise simply declare the test via `test(...)`'));
 		}
 
-		if (this.assertError !== undefined) {
-			return this.exit();
-		}
-
-		ret
-			.then(function () {
+		ret.then(
+			function () {
 				self.exit();
-			})
-			.catch(function (err) {
+			},
+			function (err) {
 				if (!(err instanceof Error)) {
 					err = new assert.AssertionError({
 						actual: err,
@@ -157,13 +158,14 @@ Test.prototype.run = function () {
 				self._setAssertError(err);
 				self.exit();
 			});
+		return this.promise().promise;
 	}
 
-	if (!this.metadata.callback) {
-		return this.exit();
+	if (this.metadata.callback) {
+		return this.promise().promise;
 	}
 
-	return this.promise().promise;
+	return this.exit();
 };
 
 Test.prototype._result = function () {
@@ -221,7 +223,11 @@ Test.prototype.exit = function () {
 	if (this.sync) {
 		self.duration = globals.now() - self._timeStart;
 		globals.clearTimeout(self._timeout);
-		return this._result();
+		var result = this._result();
+		if (this.report) {
+			this.report(result);
+		}
+		return result;
 	}
 
 	Promise.all(this.assertions)