--- a/lib/test.js
+++ b/lib/test.js
@@ -102,7 +102,7 @@ Test.prototype.promise = function () {
 	var self = this;
 	if (!this._promise) {
 		this._promise = {};
-		this._promise.promise = new Promise(function (resolve, reject) {
+		this._promise.promise = new Promise(function (resolve, reject) { // eslint-disable-line
 			self._promise.resolve = resolve;
 			self._promise.reject = reject;
 		}).tap(function (result) {