--- a/lib/test.js
+++ b/lib/test.js
@@ -307,10 +307,16 @@ PublicApi.prototype = enhanceAssert({
 Object.defineProperty(PublicApi.prototype, 'context', {
 	enumerable: true,
 	get: function () {
-		return this._test.contextRef.context;
+		var contextRef = this._test.contextRef;
+		return contextRef && contextRef.context;
 	},
 	set: function (context) {
-		this._test.contextRef.context = context;
+		var contextRef = this._test.contextRef;
+		if (!contextRef) {
+			this._test._setAssertError(new Error('t.context is not available in ' + this._test.metadata.type + ' tests'));
+			return;
+		}
+		contextRef.context = context;
 	}
 });