--- a/lib/test.js
+++ b/lib/test.js
@@ -13,7 +13,7 @@ var globals = require('./globals');
 
 function Test(title, fn, contextRef, report) {
 	if (!(this instanceof Test)) {
-		return new Test(title, fn, contextRef);
+		return new Test(title, fn, contextRef, report);
 	}
 
 	if (typeof title === 'function') {