The main premise I had while creating this: Deferring when you don't need to is expensive.

 This rewrites the runner and test to use the following semantics:

If a test never uses an async features (i.e. `cb` mode, `t.throws(promise)`, return a `promise` / `observable`) the next test is executed synchronously.

This requires a unified API between tests and the runner:

```js
// returns a "result" or promise for a result
var result = runnable.run();

// a "result" has the following properties:
result = {
  passed: true || false,
  reason: Error,
  result: "some result"
}
```

Because `run` can return either a result or a promise for a result, you frequently need to check which it is using `is-promise`. This leads to some verbose code, but it runs faster, and is fairly well encapsulated in the `sequence` and `concurrent` classes.

AVA master:
```
ava --verbose  9.38s user 0.63s system 461% cpu 2.173 total
```

This PR:
```
ava --verbose  8.46s user 0.65s system 522% cpu 1.744 total
```