@jamestalmage 2016-01-24
All working now, using benchmarks from [jamestalmage/emoji-aware#no-assertion-plans](https://github.com/jamestalmage/emoji-aware/tree/no-assertion-plans):

Master:
```
ava --verbose  6.86s user 0.56s system 439% cpu 1.689 total
ava --verbose  6.78s user 0.55s system 438% cpu 1.669 total
ava --verbose  6.73s user 0.55s system 432% cpu 1.681 total
```

This PR:
```
ava --verbose  5.25s user 0.53s system 432% cpu 1.337 total
ava --verbose  4.97s user 0.50s system 419% cpu 1.303 total
ava --verbose  5.05s user 0.50s system 426% cpu 1.301 total
```