@jamestalmage 2016-01-24
Using the following benchmark (no assertions - just tracking runner performance):

```js
import test from './';

for (var i = 0; i < 10000; i++) {
	test('test' + i, t => {});
}
```

Master:
```
node ./cli.js benchmark.js --verbose  3.15s user 0.31s system 104% cpu 3.318 total
node ./cli.js benchmark.js --verbose  2.76s user 0.28s system 103% cpu 2.937 total
node ./cli.js benchmark.js --verbose  2.73s user 0.28s system 103% cpu 2.911 total
```

This PR:
```
node ./cli.js benchmark.js --verbose  1.44s user 0.25s system 107% cpu 1.573 total
node ./cli.js benchmark.js --verbose  1.40s user 0.25s system 106% cpu 1.544 total
node ./cli.js benchmark.js --verbose  1.38s user 0.24s system 108% cpu 1.499 total
```