@jamestalmage 2016-01-24
This also seems to provide a slight performance boost for async tests:

```js
import test from './';

for (var i = 0; i < 10000; i++) {
	test('test' + i, t => {return Promise.resolve()});
}
```

Master:

```
node ./cli.js benchmark.js --verbose  2.84s user 0.28s system 103% cpu 3.010 total
node ./cli.js benchmark.js --verbose  2.86s user 0.28s system 103% cpu 3.033 total
node ./cli.js benchmark.js --verbose  2.82s user 0.28s system 104% cpu 2.980 total
```

This PR:
```
node ./cli.js benchmark.js --verbose  2.68s user 0.28s system 103% cpu 2.843 total
node ./cli.js benchmark.js --verbose  2.69s user 0.28s system 104% cpu 2.854 total
node ./cli.js benchmark.js --verbose  2.72s user 0.27s system 104% cpu 2.880 total
```