@jamestalmage 2016-01-24
Mixed sync/async:
```
import test from './';

test.beforeEach(t => {});
test.afterEach(t => {});

for (var i = 0; i < 10000; i++) {
	test('test' + i, t => {return Promise.resolve()});
}
```

Master:
```
node ./cli.js benchmark.js --verbose  4.78s user 0.43s system 102% cpu 5.075 total
node ./cli.js benchmark.js --verbose  4.64s user 0.42s system 102% cpu 4.953 total
node ./cli.js benchmark.js --verbose  4.75s user 0.43s system 102% cpu 5.041 total
```

This PR:
```
node ./cli.js benchmark.js --verbose  3.56s user 0.38s system 103% cpu 3.797 total
node ./cli.js benchmark.js --verbose  3.56s user 0.38s system 103% cpu 3.791 total
node ./cli.js benchmark.js --verbose  3.62s user 0.39s system 103% cpu 3.888 total
```