@matthewbauer 2016-01-25
Is there a way to get nested groups (#222) working in this? I'm thinking of an implementation like this:

```js
test.group(function(test) {
  test("test a", function() {
  });
  test.beforeEach("before test a but not test b", function() {
  });
});
test("test b", function() {});
```

Basically test.group just generates another runner that is added into the Runner.prototype.tests array. All of the hooks work as expected, just with the new test in context.

I was working on a PR to do this but then I saw the changes here and wanted to check what your thoughts are before moving forward.