@sindresorhus 2016-01-25
:clap: From a quick skim this looks very promising, although a lot to take in. I like the improved separation of concerns and performance improvements, but not totally happy about the added verboseness. Would be nice if we could follow up later with ways to simplify it. I'll try to do a more thorough review tomorrow.

-

@vdemedes @ariporad @sotojuan @novemberborn I could use some help reviewing this :)