@sindresorhus 2016-01-25
@matthewbauer Can you move your comment to https://github.com/sindresorhus/ava/issues/222 so not to derail this PR discussion? :)