@jamestalmage 2016-01-25
> Would be nice if we could follow up later with ways to simplify it.

Agreed. Getting a good benchmarking suite (#460) would be really helpful as well.