@vdemedes 2016-01-25
Massive applause for huge amount of work, @jamestalmage.

I have some style/cleanup-related suggestions, but I can follow up with a PR myself ;)