@sindresorhus 2016-01-25
> I have some style/cleanup-related suggestions, but I can follow up with a PR myself ;)

Yes, let's do style stuff later. (I did a few myself, but realized it's better not to dilute the review with nitpicks).