@jamestalmage 2016-01-25
Once merged, we should all dogfood this on existing projects before release, (maybe deploy with a `next` tag for a while as well). There are enough changes here that our regression exposure is pretty large. I've added pretty exhaustive tests, but we should still thoroughly vet it.