@sindresorhus 2016-01-25
> Once merged, we should all dogfood this on existing projects before release

:+1: I've already tried it on `got` and `pageres` and seems to work good.

> maybe deploy with a next tag for a while as well

:+1: