@sotojuan 2016-01-25
Currently in class but from what I skimmed it looks good :-) Awesome work!