@sindresorhus 2016-01-27
Landed! Awesome work @jamestalmage :)

![unicorn17](https://cloud.githubusercontent.com/assets/170270/12619424/3d37b2fa-c518-11e5-8d55-310953dc9767.gif)